% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_functions.R
\name{categorize_travel_time}
\alias{categorize_travel_time}
\title{Categorize Travel Times}
\usage{
categorize_travel_time(travel_time, categories)
}
\arguments{
\item{travel_time}{Numeric vector of travel times (in minutes)}

\item{categories}{List with breaks and labels (from create_categories())}
}
\value{
Factor vector with travel time categories
}
\description{
Assigns travel time values to predefined categories using cut().
}
\examples{
# Create categories
cats <- create_categories(
  breaks = c(0, 30, 60, Inf),
  labels = c("Close", "Medium", "Far")
)

# Categorize some travel times
times <- c(15, 45, 75, 25)
categorized <- categorize_travel_time(times, cats)
print(categorized)

}
