% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_population_raster}
\alias{get_population_raster}
\title{Get Kenya Population Raster}
\usage{
get_population_raster()
}
\value{
RasterLayer object
}
\description{
Returns the WorldPop 2020 population raster for Kenya
}
