% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{get_transition_matrix}
\alias{get_transition_matrix}
\title{Get Kenya Transition Matrix}
\usage{
get_transition_matrix()
}
\value{
Transition matrix object
}
\description{
Returns the motorized transport transition matrix for Kenya
}
