% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation_functions.R
\name{process_access_data}
\alias{process_access_data}
\title{Process accessibility data to data.table}
\usage{
process_access_data(access_raster, populationRaster, categories)
}
\arguments{
\item{access_raster}{RasterLayer with travel times}

\item{populationRaster}{Population raster resampled to match accessibility surface (from compute_travel_cost())}

\item{categories}{Travel time categories (from create_categories())}
}
\value{
data.table with columns: x, y, pop, travel_time, cat_travel_time
}
\description{
Converts accessibility and population rasters to a data.table with
coordinates, population counts, travel times, and travel time categories.
}
\examples{
\dontrun{
 # After computing travel costs
data <- load_default_data()
points <- validate_points(facility_coords)
results <- compute_travel_cost(points, data$transitionMatrix, data$populationRaster)
# Process to data.table
cats <- default_categories()
dt <- process_access_data(results$access, results$pop, cats)
# Examine results
head(dt)
dt[, .N, by = cat_travel_time]

}

}
